confluence-mcp/pyproject.toml
[project]
name = "confluence-mcp"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastmcp>=2.11.2",
  "httpx>=0.28.1",
  "python-dotenv>=1.0.0",
  "markdown2>=2.4.10",
  "beautifulsoup4>=4.12.2",
  "lxml>=5.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# 👇 Tell hatch which files belong in the wheel
[tool.hatch.build.targets.wheel]
packages = ["services", "utils"]          # include the services/ package
include = ["confluence.py"]      # also include the top-level script

# (optional, but tidy)
[tool.hatch.build.targets.sdist]
include = [
  "services", 
  "utils",
  "confluence.py", 
  "pyproject.toml", 
  "README.md",
  "config.yaml",
  "config.mac.yaml",
  "config.windows.yaml"
]
